Some clean-up.
