--- token_manager/src/handlers/distribution/DividendDistributionHandlers.tsx
@@ -1,3 +1,0 @@
-import { DividendDistribution } from "@polymathnetwork/polymesh-sdk/types";
-
-export type OnDividendDistributionCreated = (action: DividendDistribution) => void
